+++ src/juicebox/data/DynamicMatrixZoomData.java
@@ -46,12 +46,13 @@ public class DynamicMatrixZoomData extends MatrixZoomData {
     }
 
     @Override
-    public List<Block> getNormalizedBlocksOverlapping(int binX1, int binY1, int binX2, int binY2,
+    public List<Block> getNormalizedBlocksOverlapping(long binX1, long binY1, long binX2, long binY2,
                                                       final NormalizationType norm, boolean isImportant, boolean fillUnderDiagonal) {
-        int higherBinX1 = binX1 * scaleFactor;
-        int higherBinY1 = binY1 * scaleFactor;
-        int higherBinX2 = binX2 * scaleFactor;
-        int higherBinY2 = binY2 * scaleFactor;
+        // for V8 will be ints
+        int higherBinX1 = (int) (binX1 * scaleFactor);
+        int higherBinY1 = (int) (binY1 * scaleFactor);
+        int higherBinX2 = (int) (binX2 * scaleFactor);
+        int higherBinY2 = (int) (binY2 * scaleFactor);
         List<Block> blocksFromHigherRes = higherResZD.getNormalizedBlocksOverlapping(higherBinX1, higherBinY1, higherBinX2, higherBinY2,
                 norm, isImportant, fillUnderDiagonal);
         return createBlocksForLowerRes(blocksFromHigherRes, norm);
